Implement fallback content for RemoteFrameOwner
RemoteFrameOwner does not implement the logic for fallback content. This
means if a cross-origin navigation fails with some error, the owner
element (in this context, <object>) in the parent process does not get
notified and will not use its fallback content (instead the frame might
 show an error page).

When the <object> has fallback content, it should always use that over
the frame's error message. To support this matter, this CL implements
fallback methods in RemoteFrameOwner. Essentially,

  * When <object> creates a local frame, the corresponding frame tree
node will be marked as the type that "can" render fallback content. This
will propagte everywhere using FrameReplicationState.

  * When the provisional loading of a frame fails, RemoteFrameOwner will
notify the browser through the current proxy for navigation. The browser
then uses the parent frame to notify the renderer process that the owner
element should render its own fallback content.

  * If the <object> does not specify fallback, the navigation to error
page commits and the error page is shown.

  * When the <object> renders its own content, the remote frame stays
alive. This is a bug but not a new one; It will be fixed in future CLs.

Bug: 853140
Change-Id: Icad3934ccfc1823c0cdecd8e1223e6370ea4b3bb